+++ src/components/Post/LikeToggle.tsx
@@ -6,7 +6,7 @@ interface ILikeToggle {
   count: number;
   selected?: boolean;
   onChangeLike: (likeCount: number) => void;
-};
+}
 
 const LikeToggle: FC<ILikeToggle> = ({ count, selected, onChangeLike }) => {
   const [selectedBtn, setSelectedBtn] = useState(selected);
@@ -31,27 +31,27 @@ const LikeToggle: FC<ILikeToggle> = ({ count, selected, onChangeLike }) => {
   useEffect(() => onChangeLike(likeCount), [likeCount, onChangeLike]);
 
   return (
-      <ToggleButton
-        value="check"
-        selected={selectedBtn}
-        onChange={onChange}
-        sx={toggleStyles}
+    <ToggleButton
+      value="check"
+      selected={selectedBtn}
+      onChange={onChange}
+      sx={toggleStyles}
+    >
+      <Box sx={{ display: 'flex', strokeWidth: 2 }}>
+        <LikeSVG
+          style={{
+            stroke: selectedBtn ? selectedColor : notSelectedColor,
+            fill: selectedBtn ? selectedColor : 'none',
+          }}
+        />
+      </Box>
+      <Typography
+        variant="subtitle1"
+        sx={{ color: notSelectedColor, lineHeight: 1.1 }}
       >
-        <Box sx={{ display: 'flex', strokeWidth: 2 }}>
-          <LikeSVG
-            style={{
-              stroke: selectedBtn ? selectedColor : notSelectedColor,
-              fill: selectedBtn ? selectedColor : 'none',
-            }}
-          />
-        </Box>
-        <Typography
-          variant="subtitle1"
-          sx={{ color: notSelectedColor, lineHeight: 1.1 }}
-        >
-          {likeCount}
-        </Typography>
-      </ToggleButton>
+        {likeCount}
+      </Typography>
+    </ToggleButton>
   );
 };
